python/control_impresoras/requirements.txt
appdirs==1.4.4
argcomplete==2.0.0
Pillow==9.0.1
pyserial==3.5
python-escpos==2.2.0
pyusb==1.2.1
PyYAML==6.0
qrcode==7.3.1
six==1.16.0
websocket-client==1.3.1
